Library-backend/requirements.txt

Flask>=2.1
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
marshmallow>=3.0
marshmallow-sqlalchemy>=0.28
python-dotenv>=0.19
pytest>=7.0
requests>=2.0
SQLAlchemy>=1.4